Extract common true/false prim code
