+++ src/som/primitives/ast/true_primitives.py
@@ -1,14 +1,4 @@
-from som.primitives.primitives import Primitives
-from som.vm.globals import falseObject, trueObject
-from som.vmobjects.primitive import Primitive
-
-
-def _not(ivkbl, rcvr, args):
-    return falseObject
-
-
-def _or(ivkbl, rcvr, args):
-    return trueObject
+from som.primitives.true_primitives import TruePrimitivesBase as _Base
 
 
 def _and(ivkbl, rcvr, args):
@@ -17,11 +7,7 @@ def _and(ivkbl, rcvr, args):
     return block_method.invoke(block, [])
 
 
-class TruePrimitives(Primitives):
+TruePrimitives = _Base
 
-    def install_primitives(self):
-        self._install_instance_primitive(Primitive("not", self._universe, _not))
-        self._install_instance_primitive(Primitive("or:", self._universe, _or))
-        self._install_instance_primitive(Primitive("||", self._universe, _or))
-        # self._install_instance_primitive(Primitive("and:", self._universe, _and))
-        # self._install_instance_primitive(Primitive("&&", self._universe, _and))
+# self._install_instance_primitive(Primitive("and:", self._universe, _and))
+# self._install_instance_primitive(Primitive("&&", self._universe, _and))
